# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kampaignz API
  description: Independent Express API for Kampaignz
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /army-rulebooks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get a specific army rulebook
      responses:
        '200':
          description: Army rulebook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArmyRulebook'
        '404':
          description: Army rulebook not found
    put:
      summary: Update a specific army rulebook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArmyRulebookInput'
      responses:
        '200':
          description: Updated army rulebook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArmyRulebook'
        '404':
          description: Army rulebook not found
  /army-rulebooks/{id}/share:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    post:
      summary: Share an army rulebook with another campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                campaignId:
                  type: integer
      responses:
        '200':
          description: Successfully shared
        '400':
          description: Validation error
  /army-rulebooks/{id}/share/{campaignId}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: path
        name: campaignId
        required: true
        schema:
          type: integer
    delete:
      summary: Remove a share from a campaign
      responses:
        '200':
          description: Successfully removed share
...
  /rules/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    put:
      summary: Update a specific rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdateInput'
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Rule not found
    delete:
      summary: Delete a rule
      responses:
        '200':
          description: Successfully deleted
        '404':
          description: Rule not found
...
components:
  schemas:
...
    ArmyRulebook:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        original_campaign_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ArmyRulebookInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    Rule:
      type: object
      properties:
        id:
          type: integer
        army_rule_id:
          type: integer
          nullable: true
        campaign_id:
          type: integer
          nullable: true
        mission_id:
          type: integer
          nullable: true
        rule_category:
          type: string
        name:
          type: string
        description:
          type: string
        metadata:
          type: string
          nullable: true
        parent_rule_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RuleUpdateInput:
      type: object
      required:
        - name
        - rule_category
      properties:
        name:
          type: string
        rule_category:
          type: string
        description:
          type: string
        metadata:
          type: string
        parent_rule_id:
          type: integer
...